refactor(cli): extract child picker hint policy
